put into optimiser_base generic stuff about extension trying to probe phase space using a predictive algo
would be suitable for bayesian and adaptive
